available_spots를 엔터티에서 제거하도록 리팩토링
