+++ ch6/repository/cache.py
@@ -7,3 +7,6 @@ class Cache:
     
     def get(self, last_name):
         pass
+
+    def put(self, a, b):
+        pass
