+++ ch6/repository/session.py
@@ -1,0 +1,6 @@
+class Session:
+    def __init__(self):
+        pass
+
+    def create_query(self, sql):
+        pass
